+++ fvr_engine-client/src/renderer_v2.rs
@@ -766,9 +766,9 @@ impl RendererV2 {
 
         // Return the world coords if they are in bounds of the faux terminal.
         if x >= 0
-            && x < (self.terminal_dimensions.0 * self.tile_dimensions.0) as i32
+            && x < self.terminal_dimensions.0 * self.tile_dimensions.0
             && y >= 0
-            && y < (self.terminal_dimensions.1 * self.tile_dimensions.1) as i32
+            && y < self.terminal_dimensions.1 * self.tile_dimensions.1
         {
             Some((x, y))
         } else {
@@ -782,7 +782,7 @@ impl RendererV2 {
     pub fn screen_to_terminal_coords(&self, (x, y): ICoord) -> Option<ICoord> {
         let world = self.screen_to_world_coords((x, y))?;
 
-        Some((world.0 as i32 / self.tile_dimensions.0, world.1 as i32 / self.tile_dimensions.1))
+        Some((world.0 / self.tile_dimensions.0, world.1 / self.tile_dimensions.1))
     }
 
     //---------------------------------------------------------------------------------------------
@@ -834,22 +834,20 @@ impl RendererV2 {
         match layout {
             // Center the glyph.
             TileLayout::Center => (
-                ((self.tile_dimensions.0 as i32 - metric.width as i32) as f32 / 2.0).floor(),
-                ((self.tile_dimensions.1 as i32 - metric.height as i32) as f32 / 2.0).floor(),
+                ((self.tile_dimensions.0 - metric.width) as f32 / 2.0).floor(),
+                ((self.tile_dimensions.1 - metric.height) as f32 / 2.0).floor(),
             ),
             // Center the glyph horizontally but align with the base of the quad vertically.
             TileLayout::Floor => (
-                ((self.tile_dimensions.0 as i32 - metric.width as i32) as f32 / 2.0).floor(),
-                (self.tile_dimensions.1 as i32 - metric.height as i32) as f32,
+                ((self.tile_dimensions.0 - metric.width) as f32 / 2.0).floor(),
+                (self.tile_dimensions.1 - metric.height) as f32,
             ),
             // Adjust the glyph based on font metrics.
             TileLayout::Text => (metric.x_offset as f32, metric.y_offset as f32),
             // Adjust the glyph from the center position by an exact offset.
             TileLayout::Exact((x, y)) => (
-                (((self.tile_dimensions.0 as i32 - metric.width as i32) as f32 / 2.0) + x as f32)
-                    .floor(),
-                (((self.tile_dimensions.1 as i32 - metric.height as i32) as f32 / 2.0) + y as f32)
-                    .floor(),
+                (((self.tile_dimensions.0 - metric.width) as f32 / 2.0) + x as f32).floor(),
+                (((self.tile_dimensions.1 - metric.height) as f32 / 2.0) + y as f32).floor(),
             ),
         }
     }
